Remove specific version tags (GameLoop/CuePlayer)
Sorry for making another pull request so fast but I am unsure about something.

I would like the quarks to default to the latest version available. I've added the current version tag as it was suggested in the help file but then I found some issues and released a bug fixes version (0.1.1). When I went to install the quark it defaulted to 0.1.0 though the new version was available for manual installation.

Will these changes give me the behavior I expect?