Update directory.txt
I removed the tag from the miSCellaneous_lib entry. Although I do mark releases I became aware that the tag in this list makes no sense in this case, as I'm only uploading release versions. So it's easier for the user to always get the latest and also for you reviewers as it wouldn't have to be changed with every new release, sorry for my delayed insight.