remove CVCenter from directory.txt
In its current configuartion CVCenter will not install properly as a
quark. It depends on a tweaked version of the Conductor library, its
helpfiles need to be updated and in general I would like to rewrite it
from scratch. It should, however, work if cloned including Conductor as
a submodule from https://github.com/nuss/cvcenter.

Signed-off-by: Stefan Nussbaumer <st9fan@gmail.com>